C2LC-192: Configure ESLint to ignore dragdroptouch
Configure ESLint to ignore errors in the dragdroptouch library.

We do this by turning on create-react-app's support for ESLint
customization. This support is currently marked as "experimental" in
the documentation (https://create-react-app.dev/docs/setting-up-your-editor/#experimental-extending-the-eslint-config).

There is currently a caching issue in create-react-app that causes
previous ESLint configurations to interfere with future runs. A
workaround is to delete the 'node_modules/.cache/eslint-loader'
directory. See:

- https://github.com/facebook/create-react-app/issues/8683
- https://webpack.js.org/loaders/eslint-loader/